onClick api call for reports
